Update copyright notice in MIT license file
We don't do copyright assignment in this organozation, so the author
of each line of content (or possibly their employer, depending on contracts)
is the copyright holder; that's normally not reflected in the license
text itself though, that's in the commit history.

I considered two copyright lines, but opted for the same license on
all repos in this org, and also because two lines tends to confuse
license checkers like the one GitHub runs (licensee).